fix buffer

+++ api/files-service.js
@@ -23,10 +23,10 @@ router.post('/upload', upload.single('file'), async (req, res) => {
     }
 
     // Access the uploaded file details
-    const { originalname, mimetype, size, path, buffer } = req.file;
+    const { originalname, mimetype, size, path } = req.file;
     _path = path;
     const fileExtension = originalname.split('.').pop();
-    const { filename } = req.body;
+    const { title, description } = req.body;
 
     // Check file type
     if (!['text/plain', 'application/pdf'].includes(mimetype)) {
@@ -34,8 +34,6 @@ router.post('/upload', upload.single('file'), async (req, res) => {
       fs.unlinkSync(path);
       return res.status(400).json({ error: 'Invalid file type' });
     }
-
-    // console.log('File Details:', { originalname, mimetype, size, path });
 
     // TODO same file name handling
     // Check if a file with the same name exists
@@ -60,48 +58,70 @@ router.post('/upload', upload.single('file'), async (req, res) => {
     //   finalFileName = `${fileNameWithoutExtension}_${suffix}.${fileExtension}`;
     // }
 
+    // Log the type of buffer
+    const buffer = fs.readFileSync(path);
+
     // Upload file to Supabase Storage fileData: {path, id, fullPath}
     const { data: fileData, error: fileError } = await supabaseClient.storage
       .from('context_uploads')
-      .upload(`user_${req.user.id}/${filename}.${fileExtension}`, buffer);
+      .upload(`user_${req.user.id}/${title}.${fileExtension}`, buffer);
 
     if (fileError) {
       throw fileError;
     }
 
-    // Update the 'uploaded_docs' array in the 'profiles' table
-    const { data: profileData, error: profileError } = await supabaseClient
-      .from('profiles')
-      .upsert(
+    const fileId = fileData.id;
+    // Create an entry in the "context_relations" table
+    const { data: contextRelationData, error: contextRelationError } =
+      await supabaseClient.from('context_relations').upsert(
         [
           {
-            id: req.user.id,
-            uploaded_docs: [fileData.fullPath],
+            title,
+            description,
+            object_id: fileId,
+            embeddings_table: null,
           },
         ],
-        { onConflict: ['id'], merge: ['uploaded_docs'] }
+        {
+          onConflict: ['id'],
+          merge: ['title', 'description', 'object_id', 'embeddings_table'],
+        }
       );
 
-    if (profileError) {
-      throw profileError;
+    if (contextRelationError) {
+      throw contextRelationError;
     }
 
-    // Remove the uploaded file
-    fs.unlinkSync(path);
-
-    // TODO pass link to new uploaded file to AI service to generate embeddings
-    // TODO link new embeddings table to the user
-
-    const res = await axios.post(`${aiBackendUrl}/upload`, {
-      supbaseFullPath: fileData.fullPath,
+    const res = await axios.post(`${aiBackendUrl}/upload`, buffer, {
+      'Content-Type': mimetype,
     });
 
-    res.status(200).json({ message: 'File uploaded successfully' });
+    const { vector_store_ids } = res.data;
+
+    // update the field embeddings_table in the context_relations table
+    const { data: updateResult, error: updateError } = await supabaseClient
+      .from('context_relations')
+      .update({ vector_store_ids })
+      .eq('object_id', fileId);
+
+    if (updateError) {
+      throw updateError;
+    }
+
+    res
+      .status(200)
+      .json({ message: 'File uploaded successfully', contextRelationData });
+    // Remove the uploaded file
+    fs.unlinkSync(path);
   } catch (error) {
     console.error('Error:', error);
     res.status(500).json({ error: 'Internal server error' });
 
-    fs.unlinkSync(_path);
+    try {
+      if (_path) fs.unlinkSync(_path);
+    } catch (error) {
+      console.error('Error:', error);
+    }
   }
 });
